word/requirements.txt
# DOCX Skill Dependencies
# Phase 1: Core Infrastructure

# Core DOCX manipulation
python-docx>=1.1.0

# XML processing (usually included with Python, but listed for clarity)
lxml>=4.9.0

# Image handling (for Phase 2)
Pillow>=10.0.0

# Optional: Development dependencies (uncomment for development)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# mypy>=1.5.0
